+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Next version
 
+### 🚀 New
+
+* [#208](https://github.com/sdss/jaeger/pull/208) Add actor command `calibrations reset-offsets` to zero the alpha/beta offsets in one or more positioners.
+
 ### ✨ Improved
 
 * Improved handling of sextant controllers.